1. Added a templated function convert that converts between vector types 2. Moved world createion/floodfillity into worl-constructor as it does not belong in main any longer. 3. Fixed a few type conversion warnings. 4. Derpaherp a lot of changes regarding rendering. 5. The Game-class now has members Renderer and World, world does not know about Renderer at all. 7. Blargh code coded to long ago.

+++ coolgameyo/PathModule.cpp
@@ -4,7 +4,7 @@
 
 static float heuristic(vec3i a, vec3i b)
 {
-    return a.getDistanceFrom(b);
+    return (float)a.getDistanceFrom(b);
 }
 static float dist_between(vec3i a, vec3i b)
 {
